Fix Window serve.bat
The social plugin requires various dependencies that are difficult to
install especially on Windows.

This commit adds mkdocs.publish.yml which inherits from mkdocs.yml
and add the social plugin.
The GitHub workflow uses this new file.
So, the scripts serve and serve.bat don't require the social plugin
dependencies.